+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## 0.2.0
 - Array variations of `combine` methods
+- `createDispatch` and `createCommit` methods